Also deprecate createKeypairWithSeed
It's replaced with createKeypairWithSeed_ for now, which mentions `Maybe` in
the resulting type. This is far more appropriate, especially for a module
marked as -XTrustworthy, to avoid having any functions be partial.

It will be short lived anyway. I plan on making an 0.1.0.0 release somewhat
soonish, where these deprecated APIs will be removed.

Signed-off-by: Austin Seipp <aseipp@pobox.com>