Add better functions for manipulating keys
See the discussion in #12 - this is for the best over the long term,
as it will make sure people always use keys correctly. Deprecate the
old functions.

Unfortunately, the constructor itself cannot be marked DEPRECATED
without also marking the *type* as deprecated, which will cause
spurious warnings. This is a known and unfixed GHC bug.

Signed-off-by: Austin Seipp <aseipp@pobox.com>